Fix multiple definition issue when including the header more than once
